fix: don't convert wallet address to ETH format
Some measurements contain wallet addresses that are not valid (cannot be
converted to ETH `0x...` format). As a result, we are no longer
publishing any measurements, which means the MERidian contract is no
longer advancing the rounds.

This change reverts all address conversions. We should filter out
invalid addresses in the evaluation service.

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>