fix: wrap creation of new rounds in transaction
I am observing different number of retrieval tasks returned in the
response to the following request:

  GET /rounds/meridian/0xaaef78eaf86dcf34f275288752e892424dda9341/410

I suspect this depends on which Cloudfare enpoint I hit. When I add
`/?fresh=1` to the URL, I bypass the CDN and get a correct response
with 400 tasks.

I think this is caused when Cloudfare hits our endpoint while we are
still populating `retrieval_tasks` table for the recently started round.

In this pull request, I am wrapping that code in a new SQL transaction,
to make the creation of new round and its tasks atomic.

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>