feat: add "startEpoch" to round details
In order to map SPARK rounds to DRAND randomness values, we need to know
the epoch (block number) when the SPARK round started.

This pull request modifies the round tracker to query the start epoch
and store in the database, and the REST API to include this new field
in the round details responses.

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>