feat: timeout & retry while waiting for FEVM tx
When waiting for confirmation of the `ie.addMeasurements()` transaction,
implement a timeout of 2 minutes and retry scheme timing out after
10 minutes in total.

The goal is to prevent spark-publish from hanging indefinitely when
the RPC API or the blockchain are not able to provide any confirmations.

Based on `add_measurements_duration_ms` collected in InfluxDB, most
transactions are confirmed within 100 seconds, which is within the
2 minute timeout. There are occassional outliers going up to
500 seconds, they will be handled by the retries. We cannot afford to
wait for the confirmation too long, otherwise we will not commit
measurements before the round ends.

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>