fix: deduplicate list of clients in retrieval tasks
Recently, we discovered that there is a lot of deals using the same
payload CID `bafkqaaa`. There are more than 22k deals, but only
between 27 miners and 9 clients. As a result, when one of such deals is
picked for testing, the list of clients is very long and full of
duplicates.

In this change, I am modifying the query sampling eligible deals
to de-duplicate the list of clients in each task.

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>